Avoid redundant checks of the key returned by ets:lookup in pg.erl
Part of pg.erl in the stdlib has the following pattern:
[{Key, Value1, Value2}] = ets:lookup(Table, Key)

Because Key appears on the left side, and is already bound,
it results in an equality check between the Key
that is returned by lookup and the one that was passed to lookup.
This is pointless since they are guaranteed to match.

I verified on a small example, and this results in 3 bytecode instructions
and 19 (!) useless x86 instructions.
It can be trivially avoided by using an unbound variable such as _Key:
[{_Key, Value1, Value2}] = ets:lookup(Table, Key)

I don't expect the performance win to be significant, but it is so simple
that I think it worthwhile to land.