Fix filelib:safe_relative_path/2
The function would return a wrong path if the second argument was a link
and the path in the first argument would descend and then ascend again to
the top level.

The function would also report a path as unsafe by falsely detecting a
link loop if a link would target another link, which in turn would target
a file of the same name like bar --> foo/bar --> (foo/)foo/bar.

Co-authored-by: Björn Gustavsson <bgustavsson@gmail.com>